Corrected signed/unsigned integer comparison warnings.
The return value of the ftell() function was assigned to an unsigned
integer typed variable and then compared to -1.  Since ftell() actually
returns a 'long' assign to a 'long' instead.  This addresses
signed/unsigned integer comparison warnings.

Reviewed-on: https://gerrit.iotivity.org/gerrit/4689
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ resource/csdk/connectivity/samples/android/casample/sampleService/src/main/jni/ResourceModel.c
@@ -1633,7 +1633,7 @@ bool read_file(const char* name, char** bytes, size_t* length)
 
     FILE* file;
     char* buffer;
-    size_t fileLen;
+    long fileLen;
 
     // Open file
     file = fopen(name, "rt");
@@ -1654,7 +1654,7 @@ bool read_file(const char* name, char** bytes, size_t* length)
     }
     fseek(file, 0, SEEK_SET);
 
-    LOGI("file size: %d", fileLen);
+    LOGI("file size: %ld", fileLen);
 
     // Allocate memory
     buffer = calloc(1, sizeof(char) * fileLen + 1);
